Move the location of the latest-reviews
