+++ packages/review/src/components/shorten-list/index.ts
@@ -1,1 +1,2 @@
 export * from './popular-reviews'
+export * from './latest-reviews'
