+++ packages/review/src/components/shorten-list/latest-reviews.tsx
@@ -1,7 +1,6 @@
-import { useLatestReviews } from '../services'
-
-import { ReviewsList } from './reviews-list'
-import type { SortingType } from './sorting-context'
+import { useLatestReviews } from '../../services'
+import { ReviewsList } from '../reviews-list'
+import type { SortingType } from '../sorting-context'
 
 interface Props {
   resourceId: string
